LPD-85801 SF

+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/background/task/CTServicePublisher.java
@@ -121,7 +121,7 @@ public class CTServicePublisher<T extends CTModel<T>> {
 		CTRowUtil.copyCTRows(ctPersistence, connection, sb.toString());
 	}
 
-	private int _getPredeletedRowCount(
+	private long _getPredeletedRowCount(
 			Connection connection, String tableName, String primaryKeyName)
 		throws Exception {
 
@@ -188,7 +188,7 @@ public class CTServicePublisher<T extends CTModel<T>> {
 		}
 
 		if (_deletionCTEntries != null) {
-			int predeletedRowCount = _getPredeletedRowCount(
+			long predeletedRowCount = _getPredeletedRowCount(
 				connection, tableName, primaryKeyName);
 
 			if (predeletedRowCount != _deletionCTEntries.size()) {
